+++ GHEtool/VariableClasses/Efficiency/COP.py
@@ -1,0 +1,2 @@
+class COP:
+    pass