+++ GHEtool/VariableClasses/Efficiency/EER.py
@@ -1,0 +1,2 @@
+class EER:
+    pass