+++ GHEtool/VariableClasses/Efficiency/SEER.py
@@ -1,0 +1,2 @@
+class SEER:
+    pass